requirements.txt
-e .
aws_cdk.aws_logs
aws_cdk.aws_lambda
aws_cdk.aws_apigateway